Store destructured in files
