+++ src/store/index.js
@@ -1,0 +1,11 @@
+import state from './state'
+import mutations from './mutations'
+import getters from './getters'
+import actions from './actions'
+
+export default {
+  state,
+  mutations,
+  actions,
+  getters
+}
